tests, mild refactoring
